pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
    "mcp[cli]>=1.25.0",
    "aiomysql>=0.2.0",
    "langchain-openai>=1.1.6",
    "langchain>=1.2.0",
    "langchain-huggingface>=0.1.0",
    "sentence-transformers>=3.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "requests>=2.32.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "autogenstudio>=0.4.2.2",
    "autogen-agentchat>=0.5.7",
    "autogen-ext[openai]>=0.5.7",
    "youdotcom>=1.4.1",
    "python-a2a>=0.5.10",
    "python-dotenv>=1.2.1",
]
